Add support shared memory
This puts configuration on a shared memory.
